Rename test folder for preprocessing
