+++ helikite/tests/level0/test_init.py
@@ -1,6 +1,6 @@
 import pytest
 from helikite.classes.cleaning import Cleaner
-from helikite.tests.cleaner.mock import MockInstrument
+from helikite.tests.level0.mock import MockInstrument
 import pandas as pd
 import datetime
